+++ src/test/kotlin/racingcar/CarTest.kt
@@ -3,6 +3,7 @@ package racingcar
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import racingcar.domain.Car
 
 class CarTest {
     @Test
